Split up functional tests into operator and helm tests
