+++ opensearch-operator/functionaltests/operatortests/deploy-and-upgrade.yaml
