+++ opensearch-operator/functionaltests/operatortests/dummy_test.go
@@ -1,4 +1,4 @@
-package functionaltests
+package operatortests
 
 import (
 	"time"
